CoreMigration - moved ProfileLevels back to init.py
